chore: update the laylout
