--- guest-program/add-go/go.mod
@@ -1,5 +1,0 @@
-module go-add
-
-go 1.22.5
-
-require github.com/zkMIPS/zkm/go-runtime/zkm_runtime v0.0.0-20240910035438-ab68657c118b
