+++ guest-program/mem_alloc_vec/rust-toolchain.toml
@@ -1,0 +1,4 @@
+[toolchain]
+channel = "nightly-2023-03-06"
+targets = ["mips-unknown-linux-musl"]
+profile = "minimal"
