--- guest-program/revme/Cargo.toml
@@ -1,30 +1,0 @@
-[workspace]
-[package]
-name = "zkm-mips-elf-revme-rust"
-version = "0.1.0"
-edition = "2021"
-
-#[[bin]]
-#name = "revme"
-#path = "src/lib.rs"
-
-[dependencies]
-zkm-runtime = { git = "https://github.com/zkMIPS/zkm", branch = "main" }
-revm = { git = "https://github.com/powdr-labs/revm", branch = "serde-no-std", default-features = false, features = [ "serde" ] }
-
-#models = { path = "../models" }
-models = { git = "https://github.com/eigmax/powdr-revme", branch = "continuations", package = "models" }
-serde = { version = "1.0", default-features = false, features = ["alloc", "derive", "rc"] }
-serde_json = { version = "1.0", default-features = false, features = ["alloc"] }
-k256 = { version = "0.13.3", features = ["ecdsa"], default-features = false }
-
-# TODO can be removed once the powdr RISCV nightly is updated
-ahash = { version = "=0.8.6", default-features = false }
-
-libc = { version = "0.2" , features = ["extra_traits"] }
-
-[profile.dev]
-panic = "abort"
-
-[profile.release]
-panic = "abort"
