--- guest-program/revme/rust-toolchain.toml
@@ -1,4 +1,0 @@
-[toolchain]
-channel = "nightly-2023-03-06"
-targets = ["mips-unknown-linux-musl"]
-profile = "minimal"
