+++ sdk/src/local/stark.rs
@@ -4,7 +4,7 @@ use elf::{endian::AnyEndian, ElfBytes};
 use zkm_emulator::state::State;
 use zkm_emulator::utils::split_prog_into_segs;
 
-pub fn prove_stark(input: &ProverInput, storedir: &str, result: &mut ProverResult) {
+pub fn prove_stark(input: &ProverInput, storedir: &str, result: &mut ProverResult) -> anyhow::Result<bool> {
     let seg_path = format!("{}/segments", storedir);
     let seg_size = input.seg_size as usize;
     let file = ElfBytes::<AnyEndian>::minimal_parse(input.elf.as_slice())
@@ -19,13 +19,15 @@ pub fn prove_stark(input: &ProverInput, storedir: &str, result: &mut ProverResul
     let (_total_steps, seg_num, state) = split_prog_into_segs(state, &seg_path, "", seg_size);
     result.output_stream = state.public_values_stream.clone();
     if input.execute_only {
-        return;
+        return Ok(false);
     }
 
     if seg_num == 1 {
         let seg_file = format!("{seg_path}/{}", 0);
-        util::prove_single_seg_common(&seg_file, "", "", "")
+        util::prove_single_seg_common(&seg_file, "", "", "")?;
+        Ok(false)
     } else {
-        util::prove_multi_seg_common(&seg_path, "", "", "", storedir, seg_num, 0).unwrap()
+        util::prove_multi_seg_common(&seg_path, "", "", "", storedir, seg_num, 0)?;
+        Ok(true)
     }
 }
